Rewrite in Deno.
