+++ .github/workflows/ci.yml
@@ -9,10 +9,27 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
+    - uses: denoland/setup-deno@v1
+      with:
+        deno-version: v1.x
 
-    - name: Build
-      run: |
-        yarn install
-        yarn build
     - name: Test
-      run: yarn test
+      run: deno test -A
+
+    - name: Get tag version
+      if: startsWith(github.ref, 'refs/tags/')
+      id: get_tag_version
+      run: echo ::set-output name=TAG_VERSION::${GITHUB_REF/refs\/tags\//}
+    - uses: actions/setup-node@v2
+      with:
+        node-version: '16.x'
+        registry-url: 'https://registry.npmjs.org'
+    - name: npm build
+      run: deno run -A ./scripts/build_npm.ts ${{steps.get_tag_version.outputs.TAG_VERSION}}
+    - name: npm publish
+      if: startsWith(github.ref, 'refs/tags/')
+      env:
+        NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
+      run: |
+        cd npm
+        npm publish
