+++ .gitignore
@@ -1,8 +1,1 @@
-/node_modules
-/bin
-/dist
-/obj
-/temp
-*.js.map
-/yarn-error.log
-*.tsbuildinfo
+/npm
