--- .mocharc.yml
@@ -1,6 +1,0 @@
-require: ts-node/register
-recursive: true
-reporter: progress
-watch-extensions: ts
-timeout: 10000
-spec: src/tests/**/*.ts
