--- .npmignore
@@ -1,16 +1,0 @@
-/scripts
-/src
-.npmignore
-CHANGELOG.md
-CODE_OF_CONDUCT.md
-rollup.config.js
-tsconfig.json
-tsconfig.*.json
-yarn-error.log
-# do not include lock files in order to not restrict the typescript version
-yarn.lock
-package-lock.json
-dprint.json
-.github
-.mocharc.yml
-.vscode
