+++ .vscode/settings.json
@@ -1,3 +1,5 @@
 {
-  "typescript.tsdk": "node_modules\\typescript\\lib"
+  "deno.enable": true,
+  "deno.lint": false,
+  "deno.unstable": false
 }
