+++ README.md
@@ -2,6 +2,7 @@
 
 [![npm version](https://badge.fury.io/js/dts-minify.svg)](https://badge.fury.io/js/dts-minify)
 [![CI](https://github.com/dsherret/dts-minify/workflows/CI/badge.svg)](https://github.com/dsherret/dts-minify/actions?query=workflow%3ACI)
+[![deno doc](https://doc.deno.land/badge.svg)](https://doc.deno.land/https/deno.land/x/dts_minify/mod.ts)
 
 Minifies TypeScript declaration files (`.d.ts` files).
 
@@ -53,7 +54,7 @@ When true, it won't remove the JS docs.
 
 ```ts
 const minifiedText = minifier.minify(inputText, {
-    keepJsDocs: true, // false by default
+  keepJsDocs: true, // false by default
 });
 ```
