+++ dprint.json
@@ -1,17 +1,10 @@
 {
-  "$schema": "https://dprint.dev/schemas/v0.json",
-  "projectType": "openSource",
   "incremental": true,
   "includes": ["**/*.{ts,tsx,js,jsx,json,md}"],
-  "excludes": [
-    "dist",
-    "lib",
-    "**/node_modules",
-    "**/*-lock.json"
-  ],
+  "indentWidth": 2,
   "plugins": [
-    "https://plugins.dprint.dev/typescript-0.45.0.wasm",
-    "https://plugins.dprint.dev/json-0.10.2.wasm",
-    "https://plugins.dprint.dev/markdown-0.7.1.wasm"
+    "https://plugins.dprint.dev/typescript-0.59.2.wasm",
+    "https://plugins.dprint.dev/json-0.13.2.wasm",
+    "https://plugins.dprint.dev/markdown-0.11.2.wasm"
   ]
 }
