--- package.json
@@ -1,35 +1,0 @@
-{
-  "name": "dts-minify",
-  "version": "0.2.3",
-  "description": "Minifier for TypeScript declaration files (.d.ts).",
-  "main": "dist/dts-minify.js",
-  "author": "David Sherret",
-  "license": "MIT",
-  "repository": "git+https://github.com/dsherret/dts-minify.git",
-  "typings": "lib/dts-minify.d.ts",
-  "scripts": {
-    "build": "rimraf dist && yarn rollup && yarn build:declarations",
-    "build:declarations": "ts-node --compiler ttypescript --project scripts/tsconfig.json --transpile-only scripts/buildDeclarations.ts",
-    "format": "dprint fmt",
-    "test": "cross-env TS_NODE_COMPILER=\"ttypescript\" TS_NODE_TRANSPILE_ONLY=\"true\" mocha",
-    "test:debug": "yarn test --inspect-brk",
-    "rollup": "rollup --config"
-  },
-  "devDependencies": {
-    "@types/chai": "^4.2.18",
-    "@types/mocha": "^8.2.2",
-    "@types/node": "^15.0.2",
-    "@types/ts-nameof": "^4.2.1",
-    "chai": "^4.3.4",
-    "cross-env": "^7.0.3",
-    "mocha": "^8.4.0",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.47.0",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "ts-morph": "npm:ts-morph@5.0.0",
-    "ts-nameof": "^5.0.0",
-    "ts-node": "^9.1.1",
-    "ttypescript": "^1.5.12",
-    "typescript": "~4.2.4"
-  }
-}
