--- scripts/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../tsconfig.common.json",
-  "compilerOptions": {
-    "outDir": "../dist-scripts",
-    "rootDir": "../",
-    "noEmit": true,
-    "module": "commonjs",
-    "plugins": [{ "transform": "ts-nameof", "type": "raw" }]
-  }
-}
