--- tsconfig.common.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2015",
-    "module": "commonjs",
-    "strict": true,
-    "sourceMap": false,
-    "removeComments": false,
-    "experimentalDecorators": true,
-    "forceConsistentCasingInFileNames": true,
-    "esModuleInterop": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitReturns": true,
-    "stripInternal": true,
-    "importHelpers": false
-  },
-  "compileOnSave": false
-}
