--- tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "./tsconfig.common.json",
-  "compilerOptions": {
-    "declaration": false,
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "plugins": [{ "transform": "ts-nameof", "type": "raw" }]
-  },
-  "exclude": [
-    "./dist",
-    "./scripts",
-    "node_modules"
-  ]
-}
