--- tsconfig.rollup.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "es2015",
-    "moduleResolution": "node",
-    "removeComments": true,
-    "plugins": [{ "transform": "ts-nameof", "type": "raw" }]
-  }
-}
